Big change
The GET for codes was taking too long. setTimeout didn't want to work. Defining the object as a const didn't work. Can only imagine Axios is causing this. Move code dates to form inputs fetch by JQ and define in main post object. This avoids the need to make the call when adding the data to the DE